requirements.txt
# 无版本（Live at Head）解释：引入版本问题，会将依赖问题递增的复杂化。无版本（永远使用当前最新）可用性通过测试保证，这导致升级工作提前化，将长痛化为一次次短痛。具体参考谷歌软件工程文献
fastapi[all]
sqlalchemy
aiomysql
httpx
pydantic_settings
locust
passlib[bcrypt]
python-jose
loguru
msgpack
redis[hiredis]
# 测试
pytest-asyncio
pytest
aiosqlite